requirements.txt
# Core ML/RL libraries
#torch>=1.12.0
#numpy<2
# cuda 12.6 cudnn 9.11
#torch==2.7.1
#torchaudio==2.7.1
#torchvision==0.22.1
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
tqdm>=4.64.0

# TSP specific libraries
tsplib95>=0.7.1

# Configuration and utilities
PyYAML
scikit-learn
seaborn
plotly
dask
pyarrow
pandas
numpy
tqdm

# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
# Optional: For better performance
# torchvision>=0.13.0  # If using CNN-based approaches
# gymnasium>=0.26.0    # Alternative to gym

# Development tools (optional)
# jupyter>=1.0.0
# ipython>=8.0.0
# black>=22.0.0        # Code formatting
# pytest>=7.0.0        # Testing
